+++ packages/design-system/src/components/Form/Form.tsx
@@ -1,7 +1,9 @@
 /** @jsx jsx */
-import { jsx, Box } from 'theme-ui';
+import { jsx } from 'theme-ui';
 
-export const Form: React.FC = (props: React.PropsWithChildren<{}>) => {
-  return <Box as="form" {...props} />;
+export const Form: React.FC<React.ComponentPropsWithRef<'form'>> = (
+  props: React.PropsWithChildren<React.ComponentPropsWithRef<'form'>>
+) => {
+  return <form {...props} sx={{ variant: 'prestyled.forms.form' }} />;
 };
 export default Form;
