--- packages/design-system/src/utils/logger.ts
@@ -1,3 +1,0 @@
-import consola from 'consola';
-consola.withTag('design-system');
-consola.wrapConsole();
