+++ packages/poolbase-app/src/app/pages/add-url.tsx
@@ -7,7 +7,7 @@ import { ListItem } from '@poolbase/design-system/src/components/ListItem';
 import { PageData } from '@poolbase/common';
 
 import PageLayout from '../components/PageLayout';
-import { firestore } from '../utils/initFirebase';
+import { firestore } from '@poolbase/common';
 import { AddUrlForm } from '../components/AddUrlForm';
 
 const AddUrlPage: NextPage = () => {
