+++ packages/poolbase-common/package.json
@@ -41,6 +41,7 @@
     "prettier": "^2.0.5"
   },
   "dependencies": {
+    "consola": "^2.11.3",
     "firebase": "^7.14.2"
   }
 }
