+++ packages/poolbase-common/src/index.ts
@@ -1,0 +1,7 @@
+import { functions } from './initFirebase';
+
+export const api = {
+  updateUserProfile: functions.httpsCallable('updateUserProfile'),
+};
+
+export * from './initFirebase';
