+++ packages/poolbase-common/src/initFirebase.ts
