+++ packages/poolbase-common/src/logger.ts
@@ -1,3 +1,2 @@
 import consola from 'consola';
-consola.withTag('app');
 consola.wrapConsole();
