+++ packages/poolbase-common/tsconfig.json
@@ -9,5 +9,5 @@
     "jsx": "react"
   },
   "exclude": ["node_modules"],
-  "include": ["**/*.ts", "**/*.tsx"]
+  "include": ["index.ts", "common.d.ts", "**/*.ts", "**/*.tsx"]
 }
